backend/requirements.txt
fastapi==0.115.2
uvicorn==0.30.0
sqlalchemy==2.0.35
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
pyjwt==2.9.0
requests==2.32.3